% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_calc.R
\name{component_calc}
\alias{component_calc}
\title{Compute auxiliary metrics}
\usage{
component_calc(object, distance, min_dist = NULL)
}
\arguments{
\item{object}{object of class sf or sfc}

\item{min_dist}{numeric indicating species dispersion}
}
\description{
Compute auxiliary metrics
}
\keyword{internal}
