% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.lconnect.R
\name{is.lconnect}
\alias{is.lconnect}
\title{Test for class lconnect}
\usage{
is.lconnect(x)
}
\arguments{
\item{x}{object to test}
}
\value{
TRUE/FALSE
}
\description{
Tests if an object belongs to lconnect class
}
\details{
is.lconnect
}
\author{
Bruno Silva

Frederico Mestre
}
