% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.pimp.R
\name{is.pimp}
\alias{is.pimp}
\title{Test for class pimp}
\usage{
is.pimp(x)
}
\arguments{
\item{x}{object to test}
}
\value{
TRUE/FALSE
}
\description{
Tests if an object belongs to pimp class
}
\details{
is.pimp
}
\author{
Bruno Silva

Frederico Mestre
}
