% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.metric.r
\name{l.metric}
\alias{l.metric}
\title{Landscape connectivity metrics}
\usage{
l.metric (landscape, metric)
}
\arguments{
\item{landscape}{Landscape object produced by upload.landscape}

\item{metric}{Landscape metrics to be computed}
}
\value{
que valores a funcao retorna
}
\description{
Compute connectivity metrics

Calculates several landscape connectivity metrics
}
\details{
l.metric
}
\examples{
exemplos de aplicacao com dados fornecidos pelo package
}
\references{
artigos em que se baseia a funcao
}
\author{
Frederico Mestre
}
