% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_imp.r
\name{patch_imp}
\alias{patch_imp}
\title{Prioritization of patches according to individual contribution}
\usage{
patch_imp(landscape, metric)
}
\arguments{
\item{landscape}{lconnect object produced by upload_land()}

\item{metric}{string indicating the landscape metric to use in the 
prioritization. Must be one of "XX", "XX" .....}
}
\value{
que valores a funcao retorna
}
\description{
Prioritization of patches according to individual contribution 
to overall connectivity. Each patch is removed at a time and connectivity 
metrics are calculated without that specific patch.
}
\details{
patch_imp
}
\references{
artigos em que se baseia a funcao
}
\author{
Frederico Mestre

Bruno Silva
}
