% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.lconnect}
\alias{plot.lconnect}
\title{Plot lconnect object}
\usage{
\method{plot}{lconnect}(x, ..., key.pos)
}
\arguments{
\item{x}{lconnect object generated by upload_land()}

\item{...}{other options passed to plot().}

\item{key.pos}{TRUE/FALSE. If TRUE adds a key to the plot}
}
\value{
Nothing. Side-effect: plots graphs.
}
\description{
Plot lconnect object with clusters indicated by different colors.
Aditional arguments accepted by plot() or plot.sf() can be included.
}
\details{
plot.lconnect
}
