% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.landscape.r
\name{upload.landscape}
\alias{upload.landscape}
\title{Dsn and layer name from .shp filepath}
\usage{
upload.landscape(path, vec)
}
\arguments{
\item{path}{The full path of the landscape file}

\item{vec}{Boolean. TRUE if file is a shapefile, FALSE
if is a raster.}
}
\value{
A SpatialPolygonsDataFrame object is returned.
}
\description{
Creates a valid dsn and layer name from a .shp filepath
}
\details{
upload.landscape
}
\examples{
vec_path <- system.file("extdata/vector.shp", package = "lconnect")
rast_path <- system.file("extdata/vector.tiff", package = "lconnect")
landscape <- upload.landscape(vec_path)
landscape <- upload.landscape(rast_path, vec = FALSE)
}
\author{
Bruno Silva
}
