% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_land.r
\name{upload_land}
\alias{upload_land}
\title{Import and convert a shapefile to a lconnect object}
\usage{
upload_land(land_path, bound_path = NULL, habitat)
}
\arguments{
\item{land_path}{string, indicating the full path of the landscape shapefile.}

\item{bound_path}{string, indicating the full path of the boundary shapefile.
If NULL (default option) a convex hull will be created and used as boundary.}

\item{habitat}{vector with habitat categories. The categories can be numeric
or character.}
}
\value{
A lconnect object is returned.
}
\description{
Import and convert a shapefile to a lconnect object. Some
landscape and patch metrics which are the core of landscape connectivity
metrics are calculated. The shapefile must be projected, i.e., in planar 
coordinates and the first field must be contain the habitat categories.
}
\details{
upload_land
}
\examples{
vec_path <- system.file("extdata/vector.shp", package = "lconnect")
landscape <- upload_land(vec_path, bound_path = NULL, habitat = 1)
}
\author{
Bruno Silva

Frederico Mestre
}
