% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_imp.r
\name{patch_imp}
\alias{patch_imp}
\title{Prioritization of patches according to individual contribution}
\usage{
patch_imp(landscape, metric)
}
\arguments{
\item{landscape}{lconnect object produced by upload_land()}

\item{metric}{string indicating the landscape metric to use in the}

\item{vector_out}{TRUE/FALSE
prioritization. The current version only allows the use of IIC.}
}
\value{
Returns a vector depicting each patch's importance to overall 
connectivity.
}
\description{
Prioritization of patches according to individual contribution 
to overall connectivity. Each patch is removed at a time and connectivity 
metrics are calculated without that specific patch.
}
\details{
patch_imp
}
\references{
#' Saura, S., Pascual-Hortal, L. (2007). A new habitat availability index to integrate connectivity in landscape conservation planning: Comparison with existing indices and application to a case study. Landscape and Unrban Planning, 83(2-3):91-103.
}
\author{
Frederico Mestre

Bruno Silva
}
