% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_imp.r
\name{patch_imp}
\alias{patch_imp}
\title{Prioritization of patches according to individual contribution}
\usage{
patch_imp(landscape, metric, vector_out = F)
}
\arguments{
\item{landscape}{lconnect object produced by upload_land()}

\item{metric}{string indicating the landscape metric to use in the}

\item{vector_out}{TRUE/FALSE}
}
\value{
Returns a vector depicting each patch's importance to overall 
connectivity.
}
\description{
Prioritization of patches according to individual contribution 
to overall connectivity. Each patch is removed at a time and connectivity 
metrics are calculated without that specific patch. The current version only 
allows the use of IIC.
}
\details{
patch_imp
}
\examples{
vec_path <- system.file("extdata/vec_projected.shp", package = "lconnect")
landscape <- upload_land(vec_path, bound_path = NULL,
                        habitat = 1, min_dist = 500)
importance <- patch_imp(landscape, metric = "IIC")
plot(importance)
}
\references{
#' Saura, S., Pascual-Hortal, L. (2007). A new habitat 
availability index to integrate connectivity in landscape conservation planning: 
Comparison with existing indices and application to a case study. Landscape and
Urban Planning, 83(2-3):91-103.
}
\author{
Frederico Mestre

Bruno Silva
}
