% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lconnect.R
\name{plot.lconnect}
\alias{plot.lconnect}
\title{Plot lconnect object}
\usage{
\method{plot}{lconnect}(x, ...)
}
\arguments{
\item{x}{lconnect object generated by \code{\link{upload_land}}}

\item{...}{other options passed to \code{\link[graphics]{plot}}.}
}
\value{
Plot with clusters
}
\description{
Method of the generic \code{\link[graphics]{plot}} for lconnect
objects.
}
\details{
Plot patches with different colours representing cluster identity.
Aditional arguments accepted by \code{\link[graphics]{plot}} or
\code{\link[sf]{plot.sf}} can be included.
}
\author{
Bruno Silva

Frederico Mestre
}
