% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lconnect.R
\name{plot.lconnect}
\alias{plot.lconnect}
\title{Plot lconnect object}
\usage{
\method{plot}{lconnect}(x, ...)
}
\arguments{
\item{x}{lconnect object generated by upload_land()}

\item{...}{other options passed to plot().}
}
\value{
Nothing. Side-effect: plots graphs.
}
\description{
Plot lconnect object with clusters indicated by different colors.
Aditional arguments accepted by plot() or plot.sf() can be included.
}
\details{
plot.lconnect
}
\author{
Bruno Silva

Frederico Mestre
}
