% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lconnect.R
\name{plot.lconnect}
\alias{plot.lconnect}
\title{Plot object of class "lconnect"}
\usage{
\method{plot}{lconnect}(x, ...)
}
\arguments{
\item{x}{Object of class "lconnect" created by \code{\link{upload_land}}.}

\item{...}{Other options passed to \code{\link[graphics]{plot}} or
or \code{\link[sf]{plot.sf}}.}
}
\value{
Plot depicting patches and cluster membership (distinct colours per 
cluster).
}
\description{
Method of the generic \code{\link[graphics]{plot}} for objects 
of class "lconnect".
}
\details{
Plot patches with different colours representing cluster membership. 
Additional arguments accepted by '\code{\link[graphics]{plot}} or 
\code{\link[sf]{plot.sf}} can be included.
}
\author{
Bruno Silva

Frederico Mestre
}
