% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pimp.R
\name{plot.pimp}
\alias{plot.pimp}
\title{Plot pimp object}
\usage{
\method{plot}{pimp}(x, ..., main)
}
\arguments{
\item{x}{pimp object generated by patch_imp()}

\item{...}{other options passed to plot().}

\item{main}{character with plot title}
}
\value{
Nothing. Side-effect: plots graphs.
}
\description{
Plot pimp object patch importance indicated by different colors.
Aditional arguments accepted by plot() or plot.sf() can be included.
}
\details{
plot.pimp
}
\author{
Bruno Silva

Frederico Mestre
}
