% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pimp.R
\name{plot.pimp}
\alias{plot.pimp}
\title{Plot pimp object}
\usage{
\method{plot}{pimp}(x, ..., main)
}
\arguments{
\item{x}{Object of class "pimp" created by \code{\link{patch_imp}}.}

\item{...}{Other options passed to \code{\link[graphics]{plot}} or 
\code{\link[sf]{plot.sf}}.}

\item{main}{String with plot title.}
}
\value{
Patch importance plot.
}
\description{
Method of the generic \code{\link[graphics]{plot}} for 
objects of class "pimp".
}
\details{
Plot patch importance with percentage value per patch. This value
indicates the percentage of reduction in the connectivity metric that the 
loss of that patch represents in the landscape. Additional arguments accepted
by \code{\link[graphics]{plot}} or \code{\link[sf]{plot.sf}} can be included.
}
\author{
Bruno Silva

Frederico Mestre
}
